feat: jsdoc & helpers refactor
