+++ src/Components/TagComparisonTable/ColumnDefs/TagsColumnDefs..tsx
@@ -1,6 +1,6 @@
 import { ColDef } from "@ag-grid-community/core"
 import { TagData } from "../../../Models/TagData"
-import { getPropertyName } from "../../../utils/common"
+import { getPropertyName } from "../../../utils/helpers"
 
 export const comparisonTagsColumnDefs = (): ColDef[] => [
     {
